git-svn-id: http://svn.code.sf.net/p/cle/code@146 a4423af8-e61d-0410-8e09-fb0f2c69f8c8

+++ clerk/trunk/cle_core/cle_instance.h
@@ -47,7 +47,7 @@
 
 typedef struct objectheader
 {
-	ushort state;
+	ulong  state;
 	ushort next_state_id;
 	ushort level;
 	ushort next_property_id;
